chrome/browser should build on the x86_64 architecture.
BUG=143395
TEST=32/64-bit chrome/browser builds without errors.
TBR=mac64

Review URL: https://chromiumcodereview.appspot.com/10825447

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152572 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/ui/cocoa/bookmarks/bookmark_bar_controller.mm
@@ -233,7 +233,7 @@ void RecordAppLaunch(Profile* profile, GURL url) {
 @synthesize delegate = delegate_;
 
 - (id)initWithBrowser:(Browser*)browser
-         initialWidth:(float)initialWidth
+         initialWidth:(CGFloat)initialWidth
              delegate:(id<BookmarkBarControllerDelegate>)delegate
        resizeDelegate:(id<ViewResizer>)resizeDelegate {
   if ((self = [super initWithNibName:@"BookmarkBar"
